feat(write post): 게시글 등록 별도 페이지로 분리(#7)
