+++ week-1/src/index.js
@@ -3,12 +3,15 @@ import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
+import {RecoilRoot} from 'recoil';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
+    <RecoilRoot>
   <React.StrictMode>
     <App />
   </React.StrictMode>
+  </RecoilRoot>
 );
 
 // If you want to start measuring performance in your app, pass a function
