+++ week-1/src/pages/Home.js
@@ -1,107 +1,29 @@
-import React, { useRef, useState } from "react";
+import React from "react";
 import BulletBoard from "../components/BulletBoard";
-import { atom, useRecoilState, useRecoilValue } from "recoil";
-
-export const postsState = atom({
-    key: "postsState",
-    default: [
-        { id: 1, subject: "subject 1", content: "contents 1" },
-        { id: 1, subject: "subject 1", content: "contents 1" }],
-});      
+import { useRecoilState } from "recoil";
+import { postsState } from "../shared/Store";
+import { useNavigate } from "react-router-dom";
+import { SlPlus } from "react-icons/sl";
 
 const Home = () => {
-    console.log("re-rendering...");
     // subject, content 상태값 추가
-    const [subject, setSubject] = useState("");
-    const [content, setContent] = useState("");
     const [posts, setPosts] = useRecoilState(postsState)
-    
-    const idRef = useRef(useRecoilValue(postsState).length + 1);
+    const navigate = useNavigate();
 
-    const onSubjectChangeHandler = (e) => {
-        setSubject(e.target.value);
-    };
-    const onContentChangeHandler = (e) => {
-        setContent(e.target.value);
-    };
-
-    const onPostClick = () => {
-        console.log("post button clicked!");
-
-        const newPost = {
-            id: idRef.current,
-            subject,
-            content,
-        };
-        setPosts([...posts, newPost]);
-        idRef.current += 1;
-    };
     return (
         <>
-            <div
-                style={{
-                    width: "100%",
-                    border: "0.5px solid black",
-                    borderRadius: "10px",
-                    padding: "5px",
-                }}
-            >
-                <div>
-                    <label
-                        htmlFor="message"
-                        className="block mb-0.5 text-sm font-medium text-gray-900 dark:text-white"
-                    >
-                        Subjects
-                    </label>
-                </div>
-                <div>
-                    <textarea
-                        id="message"
-                        onChange={onSubjectChangeHandler}
-                        rows="1"
-                        className="block p-2.5 w-full text-xs text-gray-900 bg-gray-50 rounded-lg border border-gray-300 focus:ring-blue-500 focus:border-blue-500 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500"
-                        placeholder="Writes the Subjects Here."
-                    ></textarea>
-                </div>
-                <div>
-                    <label
-                        htmlFor="message"
-                        className="block mb-0.5 text-sm font-medium text-gray-900 dark:text-white"
-                    >
-                        Contents
-                    </label>
-                </div>
-                <div>
-                    <textarea
-                        id="message"
-                        onChange={onContentChangeHandler}
-                        rows="4"
-                        className="block mb-1.5 p-2.5 w-full text-xs text-gray-900 bg-gray-50 rounded-lg border border-gray-300 focus:ring-blue-500 focus:border-blue-500 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500"
-                        placeholder="Writes the Contents Here."
-                    ></textarea>
-                </div>
-                <div style={{ display: "flex", justifyContent: "flex-end" }}>
-                    <button
-                        type="button"
-                        onClick={onPostClick}
-                        className=" text-white bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:ring-blue-300 font-medium  text-sm px-4 py-1.2  dark:bg-blue-600 dark:hover:bg-blue-700 focus:outline-none dark:focus:ring-blue-800
-                disabled:bg-gray-400 disabled:cursor-not-allowed"
-                        disabled={subject.length < 10}
-                        style={{ borderRadius: "4px" }}
-                    >
-                        Post
-                    </button>
-                </div>
-            </div>
-
             {/* card 영역 */}
-            <div className="flex w-full flex-wrap-reverse justify-center gap-4 p-4">
+            <div className="flex w-full flex-wrap justify-center gap-4 p-4">
                 {posts.length > 0 ? (
                     posts.map((post) => <BulletBoard post={post} setPosts={setPosts} />)
                 ) : (
                     <div>게시물이 없습니다.</div>
                 )}
             </div>
+            {/* 글쓰기 페이지 이동 버튼 */}
+            <button onClick={()=>navigate('/write')} className="fixed bottom-5 right-5 p-3 ">
+                <SlPlus size={70}/>
+            </button>
         </>
     );
 };
