+++ week-1/src/shared/Router.js
@@ -2,6 +2,7 @@ import React from "react";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import Home from "../pages/Home";
 import Layout from "./Layout";
+import WritePost from "../pages/WritePost"
 
 const Router = () => {
   return (
@@ -9,6 +10,7 @@ const Router = () => {
      <Layout>
       <Routes>
         <Route path="/" element={<Home />} />
+        <Route path="/write" element={<WritePost />} />
       </Routes>
      </Layout>
     </BrowserRouter>
